requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.17.0
numpy>=1.24.0
supabase>=2.0.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
fpdf2>=2.7.0
